core: Add VM external data monitoring
TPM and NVRAM data is currently updated when a VM is destroyed or
powered off.  But this is not sufficient, if the host crashes or
disappears then the data may not be available.  If the data was
updated during the VM run then the updates would be lost.

To make the situation better, this patch introduces periodic TPM and
NVRAM data monitoring.  If VM stats provide a changed hash of the
data, the data is retrieved.  Vdsm may or may not, depending on data
stability, actually provide the data -- if it is not provided then
it's likely to be provided in one of the next cycles.  If data
retrieval is successful then the data is stored to the database and
its monitoring hash is updated not to ask for the same data anymore.

Change-Id: I865bebf2b66e64df865f3f6a4ae89abbbd9ff6c1
Bug-Url: https://bugzilla.redhat.com/1932485
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>