webadmin: Filter out LUN disks before cloning
LUN disks are not included when a VM is cloned. That means they can
safely be filtered out before the disk list is shown in the Clone VM
dialog. Such filtering solves the problem with DiskAllocationModel, that
is used in the dialog and assumes that all disk are represented with
DiskImage objects, so LunDisk cannot appear there.

Change-Id: I2b7da1d576a82f5e559b4614cbfa6ca6ea9fcafb
Bug-Url: https://bugzilla.redhat.com/1892525
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>