core: refactor delete VM checkpoint to not use CoCo framework
Since bug https://bugzilla.redhat.com/1901835 was fixed, the
engine doesn't need to persist the checkpoint XML anymore.

As a side effect of that, when the root checkpoint is removed,
there is no need to fetch the updated checkpoint XML for the
new root checkpoint and persist it in the database. This makes
the checkpoint deletion process to be much simpler.

This patch simplifies the implementation of checkpoint removal and
changes it to be more atomic.

Change-Id: Id82d0b94e39894d7d6349e8f210ddfb9e25f38f4
Bug-Url: https://bugzilla.redhat.com/1919248
Bug-Url: https://bugzilla.redhat.com/1897049
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>