core: remove scratch disks if VM paused
In case of insufficient space on the storage that
resides the scratch disks and the VM's disks, the VM
will enter to 'paused' mode during the backup.

If the backup will finalize while the VM in that state
the scratch disks will not be removed.

This patch fixes that behavior by removing the scratch disks
even if the VM status is 'paused'.

Change-Id: Idf1448b3be68e796bc769db45d83d09c33c70dfe
Bug-Url: https://bugzilla.redhat.com/1936185
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>