backend: Handle sriov failover in ActivateDeactivateVmNicCommand
The sriov failover device will be handled by the
ActivateDeactivateVmNicCommand. This has a significant benefit,
as direct user actions (plug/unplug) should operate with both
devices. On the other hand migration flow will operate only
with the VF. The only flow that cannot be handled by this command
is start of vm.

Change-Id: I1cd24b12a1d65a86176821429acd57a198006e70
Bug-Url: https://bugzilla.redhat.com/1688177
Signed-off-by: Ales Musil <amusil@redhat.com>