backend: Always include alias in interface metadata
Due to the limitation of failover device for VF, that
it has to have the same MAC, we need to also include
alias for the metadata so vdsm is able to distinguish
between them.

Previously, the XML contained only MAC address of specified
interface. This is backwards compatible as vdsm is still
matching based on MAC address only exception are sriov
VF and failover device.

Change-Id: I540fd3400d5edcf36a518ae1ae9e95f21eea3692
Bug-Url: https://bugzilla.redhat.com/1688177
Signed-off-by: Ales Musil <amusil@redhat.com>