core: adjust command for changing disk to use PDIV
Currently, it's not possible to use CDs which are on block SD. To be
able to use it, volume representing CD has to be specified by PDIV.
Support for it ws added into the vdsm and host with recent vdsm reports
that it's able to accept CD specified by PDIV by setting
`cd_change_pdiv` in capabilities to `True`. In such case CD is
specified via PDIV, otherwise we fall back to old imlpementation where
CD is specified using its path.

Change-Id: I09545410333952bf5d8e913eaf4676c7f2742bb2
Bug-Url: https://bugzilla.redhat.com/1589763
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>