core: add a custom property for extra cpu flags
We see that there are some CPU flags that users may want to set
depending on the workloads that run within their virtual machine,
like CPU features intended for machine learning workloads.

The fact that those features should be set on a per-VM basis,
rules out the option of adding them to our CPU types.

Up until now, we advised users to use CPU host-model in order
to leverage such features. The downside is that the VM then
needs to be pinned to hosts whose CPU supports those features.
The VM might move to a different host and then "lose" those
features.

This patch adds a new custom property named 'extra_cpu_flags'
that adds additional flags to those that are defined for the
selected CPU type.

Not only that it allows users to set such features on a per-VM
basis and the VM won't start without them, but it also simplifies
testing of CPU features - one can just add the given flag with
this custom property and check that the VM starts well (otherwise,
one needs to inspect the configuration within the guest).

Change-Id: I7f5f8cf5a0e0b66c80e941818a33748c22e11a91
Signed-off-by: Arik Hadas <ahadas@redhat.com>