core: fix VM and VM template import with old stale configuration
Fixing an issue with importing a VM / VM template due to a
stale data that exists in "vm_ovf_generations" DB table.

The issue was that when attaching the SD, the
"unregistered_ovf_of_entities" DB table was populated
from the existing (and stale!) configuration
taken from the "vm_ovf_generations" DB table.

The solution is, when executing SD Attach and populating the
"unregistered_ovf_of_entities", to use the configuration taken
from the "vm_ovf_generations" DB table only when no configuration
is provided ("ovf_data" was passed as null).
Otherwise, the flow should take the provided up-to-date configuration.

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1910858
Change-Id: I3a88cb23f8fe5d2c513e3c908c94a4f4ea4d7789