core: Add VM external data to snapshot OVF
When a VM with TPM or secure boot data is restored from a snapshot,
the corresponding TPM/NVRAM data must be restored as well.  This patch
adds support for writing the data to and reading it from OVF when
making snapshots.

Data hashes are not stored into OVF.  The primary purpose of storing
the hashes in the database is to avoid external data retrieval for all
the running VMs on Engine restart.  But that's not applicable here and
the data (and its hash) is likely to change slightly on VM start
anyway.

Change-Id: Ib82eef44319885c3a06bc1e04fc56fd8f44d57a2
Bug-Url: https://bugzilla.redhat.com/1932485
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>