core: add managed field to iSCSI MBS disks
When running VMs with MBS disks, add a "managed" attribute when the
disk is an iSCSI MBS disk. This is needed to distinguish between direct
LUNs and MBS disks, as there are different code paths required in vdsm.

For example, when running a VM, we attach the MBS disks before starting
the VMs, we also want to prepare the disks for usage. As there are
multiple flows requiring preparation and attachment (copy/move disk), we
now need to move preparation out of the start VM flow, and into
the shared attachment flow.

Bug-Url: https://bugzilla.redhat.com/1906074
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>
Change-Id: I2eb9c5b7a55b84dbb2082280357b53a3dd1f4c23