webadmin: display type selection for blank template
When editing the Blank template, the display type
was not shown correctly according to the db value,
cirrus was pre-selected instead. However, the graphic
type was selected correctly according to the db value,
which resulted in inconsistency. Also, the USB field
was always deselected as a consequence of selecting
the cirrus.

The reason for this was the way the items were inserted
into the ListModel.
Using setItems(Collection<T> value, T selectedItem)
results in invoking the selection model that synchronizes
its selected value back with the ListModel.
The scenario went as follows:
1. The items were set to display model and the default
cirrus value was selected. The selection model was invoked
and a synchronization method scheduled.
2. The correct value was selected (qxl) using setSelectedItem
method - this does not invoke the selection model.
3. The selection model's synchronization method was invoked
and changed the value from qxl to cirrus.

Change-Id: I277c61ebc2983cecea71cab316915718286f10cc
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>