core: Fix reported memory size after memory hot plug
The reported new memory size in engine.log is the same as the old
memory size.  This is because it's taken from the VM memory value in
parameters, which is the old value.  The right thing to get from the
parameters is the amount of memory hot plugged and we can use it to
compute the new memory value.

Change-Id: I33f4e3677f181e25e1b168bf692d6250e6b59232
Bug-Url: https://bugzilla.redhat.com/1775145
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>