core: prevent races between devices monitoring and attach/hot-plug disk
This patch introduces a solution for the following race:
1. The VM devices monitoring retrieves VM configuration V1
2. New disk is plugged to the running VM
3. The engine persists VM configuration V2 with the disk plugged.
4. the engine processes V1 that doesn't include the plugged disk and
thus unplug the disk.

It's not so likely that we will end up with an unplugged disk but even if
immediately after step 4 the disk is set as plugged again, there is a
possibility that another attach/hot-plug disk operation will take place
when the disk is set as unplugged at step 4 and as a result the new disk
will be assigned with the same address that was assigned to the first disk.

The solution is to add a lock that guarantees the following:
1. VM devices monitoring won't override updates that the attach, detach
and hot-plug disk commands do to disk devices.
2. The attach, detach and hot-plug disk operations are done sequentially
per VM. This renders the engine lock that AbstractDiskVmCommand used to
acquire when allocating PCI addresses redundant (and thus it is removed).

Change-Id: I1292574d3dd32271249d83d1dddb5afabe3672fe
Bug-Url: https://bugzilla.redhat.com/1927243
Signed-off-by: Arik Hadas <ahadas@redhat.com>