packaging: setup: Initialize env[SELINUX_BOOLEANS]
It's already initialized in another plugin, but this plugin was recently
moved from -common to the engine plugin. So on "separate machines",
where the engine plugin is not installed, it's missing, thus failing
setup.

This patch is probably the minimal possible solution, initializing it in
the only plugin in -common that uses SELINUX_BOOLEANS. In theory it
might have made sense to do this elsewhere, or even partially reverting
the said recent move - moving the plugin back to -common and just move
to the engine the bit we wanted to move. I think it's good enough as-is.

Change-Id: I5cecf489bc3bb9de1bb628fe8ce4d9cb8a8e3b36
Bug-Url: https://bugzilla.redhat.com/1946502
Signed-off-by: Yedidyah Bar David <didi@redhat.com>