core: drop VmDeviceDao#updateHotPlugDisk
Now that we execute the code that updates the hot-plugged disk
device in HotPlugDiskToVmCommand after acquiring a lock that
prevents the devices monitoring from updating the disk device
at the same time, we no longer need to call 'updateHotPlugDisk'
and can rather just call VmDeviceDao#update.

This patch also includes cleanup in HotPlugDiskToVmCommand.

Change-Id: Ibbb03fec79ee408d63c26e968a26974bf0d706b3
Bug-Url: https://bugzilla.redhat.com/1927243
Signed-off-by: Arik Hadas <ahadas@redhat.com>