core: reduce num of db operations on hot plug disk
Previously, AbstractDiskVmCommand#updateDeviceAddress used to lock
the VM with LockingGroup.VM_DISK_HOT_PLUG, assign an address to the
hot-plugged disk and persist the address in the database to avoid
having another hot-plug disk command from assigning the same address
before VM devices monitoring updates the address of the plugged disk.

However, now that the execution of both the code that assigns the
address and the code that persists it is done after acquiring a lock
that prevents other commands from assigning addresses to other disks
of the same VM, we can drop the code that persists the address to
the database in updateDeviceAddress and persist it later on in the
flow of HotPlugDiskToVmCommand (and AttachDiskToVmCommand).

Change-Id: Ib6a8a20762bfcd71cad4ab1101c757e16c7c02b1
Bug-Url: https://bugzilla.redhat.com/1927243
Signed-off-by: Arik Hadas <ahadas@redhat.com>