core: change balloon behvaior
Balloon device is needed in order to take memory statistics of the VM.
We need it always to exist, but we don't always want ballooning
functionality. We now always add a ballooning device, while
we add a separate indicator for whether the balloon functionality is
needed. The indicator is inserted to the VM metadata to be consumed
by VDSM.

Change-Id: Ia8a3df0eaa495a28d20103efa5f3b662119c420f
Bug-Url: https://bugzilla.redhat.com/1917718
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>