core: command-level lock for import-vm-from-configuration
ImportVmFromConfiguration acquires the same locks as ImportVm and also
uses the same callback, however, it holds the locks only for the duration
of the 'execute' phase. This leaves the VM unlocked between the end of
the 'execute' phase and the 'end-action' phase. In that period of time
the VM can be started. If it's started, ImportVmFromConfiguration would
later changes its status to 'Down' when it "unlocks" the VM at the end
of its execution.

This patch changes ImportVmFromConfiguration to hold its locks for the
entire duration of its execution (i.e., including its asynchronous phase),
as ImportVm does, to prevent the VM from being started before this
command ends.

Change-Id: I61e972ef44684b1f69a7dde40fac772deddb5400
Bug-Url: https://bugzilla.redhat.com/1917956
Signed-off-by: Arik Hadas <ahadas@redhat.com>