core: Use tpm-crb TPM model on x86
Libvirt selects a default TPM model if it is not specified.  The
default model selected by libvirt is tpm-tis on x86_64, regardless of
the selected TPM version.  But with TPM 2.0, we can use tpm-crb model,
which is a newer TPM interface.  Let's select it on x86.

On POWER, there is only tpm-spapr model available, which is selected
by libvirt automatically and no change is needed.

Change-Id: I726d91fccbec90cb7d423566ab2f1822874c98ae
Bug-Url: https://bugzilla.redhat.com/1948500
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>