packaging: backup: Fix grafana backup
The database file might change during backup and tar then fails, saying
that the file was changed while reading it.

We already handled this [1], by backing up /var/lib/grafana/*
excluding the db, and backing up the db using sqlite, but forgot to also
remove /var/lib/grafana itself from the list of paths to back up. Do
this now.

When trying to reproduce this issue, and verify the current patch,
I realized that the existing solution using sqlite .backup is also
not enough - it sometimes does fail, with "Error: database is locked".
To fix this, I now also set a timeout for locking, of 10000
milliseconds.

As an artificial test, I verified this with an endless loop of [2]
running in parallel. Without this patch, backup failed quite often.
With a timeout of 5 milliseconds, it failed very rarely, and with 100
it never failed (after several thousand attempts). This is artificial,
obviously - I have no idea how long might actual grafana transactions
take. I think 10 seconds should be ok for most cases.

[1] https://bugzilla.redhat.com/1879373
commit 80c2782, change id I414a0faa

[2] grafana-cli admin reset-admin-password somepassword

Change-Id: Ibf9f61fe5f853f4dfd483e077e0c79ac2037b65d
Signed-off-by: Yedidyah Bar David <didi@redhat.com>