restapi: Use codehaus JAXB2 maven plugin
This patch migrates a maven plugin: org.jvnet.jaxb2.maven2:maven-jaxb2-plugin[1]
used for generating of restapi model into mainstream jaxb2-maven-plugin[2].
The motivation is to use well maintained plugin in place of the one already outdated,
which according to documentation supports up to java 9.
Another reason for the migration is the initiative to use xmvn to build ovirt-engine
and rely on 3rd party dependencies being provisioned via RPM packages for downstream builds.

[1] https://github.com/highsource/maven-jaxb2-plugin
[2] https://www.mojohaus.org/jaxb2-maven-plugin/Documentation/v2.5.0/index.html

Signed-off-by: Artur Socha <asocha@redhat.com>
Change-Id: I69ccd562e6a3c1f0f1249b3cb295bac85ee17f5b