ansible: Create client migration certificates
The libvirt version in Advanced Virt 8.4 requires now client (source)
certificates in addition to server (destination) certificates for
encrypted migrations.  If a client certificate is not provided, an
encrypted migration fails with "Cannot write to TLS channel:
Input/output error".

We can reuse server certificates as client certificates, the primary
thing libvirt cares about is whether the client certificate is signed
by a proper certification authority.  This patch adds the
corresponding links from server certificates to client certificates to
make encrypted migrations to 8.4 destinations working.

This patch doesn't solve all the problems though.  The migrations fail
for source hosts that don't have the client certificates.  That means
for all hosts not having certificates installed by Engine that already
includes this patch.  Certificates must be re-enrolled for such hosts.
But that requires switching those hosts to the maintenance, which may
not be possible in case they are running VMs that cannot be migrated
to hosts running Advanced Virt < 8.4, e.g. when there is only a single
host running an old Advanced Virt.

The possible solutions, not addressed by this patch, are:

- Switching off the VMs and restarting them elsewhere.
- To use non-encrypted migrations to move the VMs to newer hosts.
- To create the certificate links manually on the old hosts.
- To set migrate_tls_x509_verify libvirt option to 0 for 8.4+ hosts
  until the cluster level is switched to 4.6 (but that would disable
  server/destination verification during migrations too).
- To prepare upgrade to 8.4 carefully to avoid the situation.

Change-Id: Ide7243d890b77e0d9fbaffc5b04c864edff90fd6
Bug-Url: https://bugzilla.redhat.com/1948376
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>