webadmin: null check in cluster column
ClusterAdditionalStatusColumn and ClusterUpgradeStatusColumn
expected the cluster's property clusterHostsAndVms
to be not null. The value is filled out correctly
when using normal search query, but
during import of a vm or template, different query
GetClustersByStoragePoolId is used to set items to
ClusterListModel. As a result, an error box was displayed to
the user during import if the clusters list screen had
been displayed before.

Change-Id: Ia600a81c073f5a60bf9fd20500a5831963a19a96
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>