Host Nic Bond Mode - improve mapping
This patch changes the way that a bond-modes description
is inferred from the bond-mode ordinal number. Until now
the description was taken from a static array in HostNicMapper.
After this patch, it will be taken from the existing 'BondMode'
enum in the Engine.

This patch isn't the final solution, though. In the future,
BondMode enum will be added to ovirt-engine-api-model, reflecting
the values of the BondMode Engine enum. This will be accompanied
potentially by some refactoring around how bond-modes are expressed
in the API.

Signed-off-by: Ori Liel <oliel@redhat.com>
Change-Id: Ia5afe2e7e9fc962dad1084880f97743f1611415b