core: block VM backup if scratch disks on shared storage not supported
If the host that runs the VM is using an old VDSM version (less than
4.4.5) and the engine version > 4.4.4 then starting a VM backup is not
possible.

In 4.4.5 version there was a major change in the way checkpoints are
redefined, the checkpoint XML generated in VDSM and not preserved in the
engine DB, also, backup scratch disks created on shared storage and
not on the host local storage, so when a backup will start on a host
with version less than 4.4.5 it will fail.

This patch adds validation for the host version, the validation
will check if the host supports the cold backup feature that was
introduced in 4.4.5 version together with the checkpoint XML
and creating scratch disks on shared storage changes.

Change-Id: I7a8a4e5265eec7840c301484c338462e72e7d113
Bug-Url: https://bugzilla.redhat.com/1950467
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>