engine: clarify error message
When updating cluster's cpu type, the error message
displayed to the user in case the host was not compatible
with the new cpu was not clear. It mentioned "lower" and
"higher" CPUs, but that naming is no longer relevant. The error
message was simply changed to state, that the host does
not support the new cpu.

Change-Id: Ie20db20edc03c8478e4ab496c0ffa7f3f1fa49ad
Bug-Url: https://bugzilla.redhat.com/1891851
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>