core: lock when adding external network
Adding an external network should behave similarly to adding a physical
network by applying a lock at the start of execution.

The AddExternalNetworkCommand which is a subclass of AddNetworkCommand
accidentally bypassed the lock in AddNetworkCommand when over-riding the
lock application method.

This patch removes the child method so that the same lock will be
applied in both commands.
This fixes the bug of being able to create two external networks with
the same name by firing two almost simultaneous AddExternalNetwork
commands.

Change-Id: I8a88f6d807723dadbee58e714b89da694bbf2ec3
Bug-Url: https://bugzilla.redhat.com/1915207
Signed-off-by: Eitan Raviv <eraviv@redhat.com>