core: filter out floating direct LUNs on Refresh LUN
In case of refreshing a floating direct LUN, or a direct LUN that
attached to a non-running VM, there's no need to sync the host,
and the host id is set to null.
In that case, running the getDeviceList command with a null host id
results with an exception on the engine log [1].

This patch filters out the floating direct LUNs before sync occurs.

[1]
EngineException: No host was found to perform the operation (Failed with error RESOURCE_MANAGER_VDS_NOT_FOUND and code 5004)

Change-Id: I90e16fb7cf16b04b508e8f99cce7ab6f8ce3989a
Bug-Url: https://bugzilla.redhat.com/1939198
Signed-off-by: Shani Leviim <sleviim@redhat.com>