ansible: Add a role for adding client migration certificate links
Encrypted migrations to RHEL 8.4 hosts require presence of client
certificates.  An ansible support for adding the corresponding links
has already been added.  But it requires the host to be on
maintenance.  If it is not possible to switch it to maintenance,
e.g. because VMs are running there and there are no hosts accepting
incoming migrations without the certificates, then this role can be
used to create the required links on running hosts.

Change-Id: I2c4ea80dc22e54a85e0eaf2bbddf6241509e93be
Bug-Url: https://bugzilla.redhat.com/1952075
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>