core: set certs on host maintenance
In case we have encrypted migration VMs that runs on a host with cluster
level lower than 4.6, the host won't be able to migrate those VMs.
Therefore, it will be impossible to get into maintenance mode; harming
upgrade flows. This patch will call ansible to set the right
certifications and allow these migrations.

Change-Id: Idd5d0d9675991e62908aff255c16f213b9878449
Bug-Url: https://bugzilla.redhat.com/1952075
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>