ansible: Add upgrade check for migration client certificate presence
Migration client certificates were added in commit
70fc2c0db733e125e88afe9895f4596956f1ab3f.  But in order to make the
certificate on host upgrade, the check for its presence must be added
to the host upgrade flow, otherwise the certificate creation will be
skipped.

This patch fixes that omission.  Note that we cannot check for the
certificate validity similar to the other certificate checks there
since that would fail if the certificate doesn't exist at all.

Change-Id: Ib6ce8b653e31c386eb297799519b2d2d3fdc7597
Bug-Url: https://bugzilla.redhat.com/1952075
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>