webadmin: perserve current pinning configuration
While we default to set `Existing` auto pinning policy to a VM when it
has high performance profile and pinned to a host, if the profile is
already selected we shouldn't re-set the pinning. That way once the
profile changes we will set the `Existing` policy defaults and on
regular cases we allow the user set is custom pinning without overriding
it on each VM update. Which causes both waste of backend work, wiping
existing configuration without user knowledge and might cause
performance issue depending on the whole configuration.

Change-Id: Ibabedee37f0b81154471f2d516cc4fed2f581968
Bug-Url: https://bugzilla.redhat.com/1954401
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>