engine: return id of last processed event
When processing events, some events are not listed as they are not
real events [1]. This causes the total number of events in the set to be
smaller than the corresponding number of the last processed event.
Thus, using 'total events' as the indicator to the number of the last
event that was processed, caused a re-run of several events, that were
listed between total events and the last event id counter.
In order to fix that, the counter, which is the first part in each event's
name is returned.

[1] https://github.com/ansible/ansible-runner/issues/656

Change-Id: Ia12cff5ce7ec7890c479923aa805916ee595c3a0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1917707
Signed-off-by: Dana Elfassy <delfassy@redhat.com>