core: Make sure async snapshots are used on newer hosts
Vdsm hosts >= 4.4 support asynchronous and only asynchronous
snapshots.  Older hosts support only synchronous snapshots.  The
decision on the Engine side whether to perform a synchronous or
asynchronous snapshot is based solely on the cluster level.  This is a
problem if a host >= 4.4 runs in a cluster < 4.4.  In such a case, a
synchronous snapshot is attempted, without providing the snapshot
timeout, making it effectively 0.  This typically results in an
aborted snapshot (due to the 0 timeout) while Engine believes the
snapshot has been successful (because the supposedly synchronous
snapshot API call, actually just starting the snapshot job, succeeds).
Less or more disastrous consequences may follow.

This patch changes the condition for asynchronous snapshot.  If the
cluster version is < 4.4 then the host is additionally checked for 4.4
cluster support.  If it is present then asynchronous snapshot is used.

Change-Id: Ia255a53da38e7a26aeb8b015956a419af9a5a2df
Bug-Url: https://bugzilla.redhat.com/1950209
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>