core: fix snapshot initial failure
If we create async snapshot, it uses the jobs infrastructure. As a
result when calling to VDSM we expect to get succeeded response and
then using the job callback to check the status of it. But, if it fails
before starting the job we will get a snapshot error response from
VDSM. We should not let the poller start in such case, where the
command will persist and wait for never happening job in VDSM.

Now, in such case we will log it and report it as a failure without
continuing to the callback. If the job in VDSM encounters an error
after the job is started the engine initially gets good response and
the callback will decide the command's success.

Change-Id: I956b85f8803ab1e7c21abfdca3de89d7337d0c6d
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>