core: sort fetched commandIds
syncChildCommandList currently relies on parameters.getChildCommands()
to have the correct order of child commands, according to their order of
execution. However, this order may change if the paramters are reloaded
from the database, which will lead to an incorrect commandStep -> commandId
mapping.

In the bug this results in fetching the ActionReturnValue for
MergeCommand instead of MergeStatusCommand, leading to missing a
MergeStatusReturnValue field.

This patch sorts the command IDs by their creation date to ensure the
latest command will be put in the list. This solution is not perfect,
and a better solution would be to have a more sophisticated mapping of
commandStep -> ActionType, but this involves greater risk as well.

Bug-Url: https://bugzilla.redhat.com/1956677
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>
Change-Id: I4d799760fd664924fcd7eca843e30bcc2d6a11ed