core: allow non-root checkpoint removal
Until now, only the root checkpoint of the VM checkpoints chain could be
deleted. So in order to remove a specific non-root checkpoint,
all checkpoints preceding this checkpoint should be removed.

Since libvirt changed the way checkpoints are handled and keep all
the checkpoints (and their bitmaps) active when the VM is up, we can
safely remove non-root checkpoints from the chain.

Change-Id: I74f41867cf5fb9098ffe29cce4f05eb621f26b3b
Bug-Url: https://bugzilla.redhat.com/1953468
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>