core: enable Manage Block Storage by default
Managed block storage can be used now by default in 4.6 clusters.
Hosts that added to a 4.6 cluster fully support all the
Managed block storage features so for those clusters it can
be used by default.

For clusters with versions lower than 4.6, the
'ManagedBlockDomainSupported' configuration value in
the database should be set to 'true' explicitly.

Change-Id: Iede72214e53d6b9f13206e710c7f2109ce1b72a1
Bug-Url: https://bugzilla.redhat.com/1957253
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>