core: stop relying on cached parameters in CreateOva
CreateOvaCommand passes the disks from its parameters to OvfOvaWriter that
tries to access their disk-vm-elements. But BaseDisk#getDiskVmElements is
set with @JsonIgnore, thus the disk-vm-elements are not persisted with the
parameters to the database. This patch loads the disk-vm-elements from the
database to prevent CreateOvaCommand from failing with an NPE when its
parameters are restored from the database (e.g., when the cache of the
command coordination framework is cleared).

Change-Id: Ibd0fa079dab73b02ab73b79f965bd5170ec61c3f
Signed-off-by: Arik Hadas <ahadas@redhat.com>