packaging: setup: pki: Allow renewing a remote cert
Before this patch, a key was generated iff its file didn't exist,
and a cert was generated iff its file didn't exist. This means that if a
cert needs to be renewed, this code does nothing.

With this patch:
- If a cert file does not exist, generate a new cert.
- If a cert file exists and is to be renewed, generate a new cert.
- If a key file does not exist, generate a new key and a new cert.
- If a key file exists, and we need to generate a cert, use the existing
key for the new cert.

This patch also moves some private functions from an existing plugin
module to a library module to be used by both this plugin and the code
for handling a remote engine.

Bug-Url: https://bugzilla.redhat.com/1959839
Change-Id: I0c3014c55479657d3db3ee569a0d100ecdb30a94
Signed-off-by: Yedidyah Bar David <didi@redhat.com>