core: update vm vnuma list
The update command calculates the differences between the given VM in
the parameters and the properties we currently have in the DB. The
vNumaNodeList doesn't appears until we query it from the NUMA DB. Until
now we only set it once we weren't asked to update that field from the
given parameters in order to validate the configuration. But, we also
need it in the current VM to prevent creating next-run configuration due
to different list compared to the VM from the databse.

Change-Id: I904f95358fce700f59c8fc12a201fa91fd07de45
Bug-Url: https://bugzilla.redhat.com/1956967
Bug-Url: https://bugzilla.redhat.com/1929730
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>