rest-api: set min-memory = memory when not specified
We usually set VM fields according to their corresponding fields
in the template when they are not specified on add VM. However,
the memory fields need to be handled differently. We already set
the max-memory, when it's not specified, according to the memory
that is specified. This patch does something similar to that for
the min-memory (minimum allocated memory):
When the min-memory is not specified and memory is specified, we
set min-memory=memory. The reason for this is that we shouldn't
assume memory overcommit by default. It makes more sense to set
min-memory=memory in the above mentioned case. If users do wish
to get memory overcommit, they should specify the min-memory
accordingly.

Change-Id: Ic1da7e016c34db061928a74ffbed178b566bdbcf
Bug-Url: https://bugzilla.redhat.com/1940529
Signed-off-by: Arik Hadas <ahadas@redhat.com>