branding: add a preamble template
Extend branding to add a `welcome_preamble` template that is inserted
on the welcome page before the main welcome template.  Only one preamble
is inserted.  The preamble in the highest ranked brand wins.

The `welcome_preamble` template goes through the same string replacement
as the main `welcome` template.  This would allow strings in the
preamble to be localized.

The default ovirt branding does not include a preamble.

Bug-Url: https://bugzilla.redhat.com/1804774
Signed-Off-By: Scott J Dickerson <sdickers@redhat.com>
Change-Id: I2559177266729d09524c2668cf6d66a778d5c14c