core: fix execution monitoring of AddVmTemplate
AddVmTemplate is monitored by a callback (on non-kubevirt clusters).

Previously, when not having cinder disks, CreateAllTemplateDisks invoked
CreateImageTemplate command(s) without setting their end-procedure and
without having a callback.
That could lead to the following situation:
1. The CreateImageTemplate command is finished
2. Its parent command CreateAllTemplateDisks is finished
3. The CreateAllTemplateDisks was thus removed from the cache
4. The callback of AddVmTemplate tried to check the status of its child
   command, but didn't find the CreateAllTemplateDisks command in the
   cache, ending up saying that the status of the child commands is
   UNKNOWN which results in failing the AddVmTemplate command although
   everything finished successfully.

Now, CreateAllTemplateDisks sets the end-procedure of each of its child
commands to EndProcedure.COMMAND_MANAGED and monitors them with a
callback, also when no cinder disks are used.
This prevents the above mentioned race since the CreateAllTemplateDisks
command is not be removed from the cache but remains there until the
callback of AddVmTemplate retrieves it.

Change-Id: I87351452a6d4ab2da97deaf88c654f9fc24e5514
Bug-Url: https://bugzilla.redhat.com/1588100
Signed-off-by: Arik Hadas <ahadas@redhat.com>