core+webadmin: block auto pinnning existing
The current behavior of auto pinning existing policy is to pin in the
best effort way the current VM CPU topology. With the current logic we
always pin it to the first physical CPUs. Unfortunetly when executing
the policy on multiple VMs on the the host, it may cause more harm then
use.

This patch blocks the option to use the existing policy. It is removed
from the UI and blocked in the backend.

Change-Id: I21b04d3a1ae27a38d7864f8f01836330e9134e78
Bug-Url: https://bugzilla.redhat.com/1963680
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>