core: add checkpoint states
Add a 'state' field for a checkpoint. The new 'state' field will
indicate the current state of the checkpoint.

Checkpoint can have the following states:

  - CREATED - when the checkpoint created and valid.
  - INVALID - when the checkpoint cannot be used anymore and should be
              removed.

The INVALID state will be used when there is a possibility for backup
corruption. It may occur when an older VM snapshot committed and the
engine removes the snapshot that was taken after the removed snapshot,
those removed snapshots may contain VM checkpoints and by removing the
checkpoints the checkpoint chain can now be corrupted.

For example, VM has the following snapshots and checkpoints chain:

snapshot-1 (contains checkpoint-1 parent:null) ->
snapshot-2 (contains checkpoint-2 parent:checkpoint-1) ->
snapshot-3 (contains checkpoint-3 parent:checkpoint-2) ->
snapshot-4 (contains checkpoint-4 parent:checkpoint-3) ->

When snapshot-2 previewed and committed, snapshot-3 and snapshot-4
removed with their checkpoints, so now those checkpoints are invalid
and the chain cannot be used.

Currently, the engine cannot remove those checkpoints during snapshot
commit since the VM is not running and VDSM is missing the
functionality to remove the checkpoint bitmap in that case.
It will be implemented in future patches.

This patch adds the new 'state' attribute to the VM checkpoint entity and
invalidate the VM checkpoints when a snapshot is committed.
Invalid checkpoint cannot be used for incremental backup, so after the
user select to commit to a snapshot a full backup should be taken.

Change-Id: I02a8bd79c31afbb8f5d9b5aef6a21317b3eda85b
Bug-Url: https://bugzilla.redhat.com/1952577
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>