core: allow removing a VM checkpoint for non-running VMs
Until now, in order to remove a VM checkpoint, the VM should
be running and the operation was done using libvirt.

This patch adds the option to remove a VM checkpoint while
the VM isn't running. In this case, the bitmap will be
removed directly from all the volumes of the image and
the checkpoint will be removed from the engine database.

Change-Id: Ib2a22135b8595e25f22606f28ff090e0c7441785
Bug-Url: https://bugzilla.redhat.com/1966177
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>