webadmin: fix selectedItemsChangedEvent
ListModelMultipleSelectListBox stores the selected items internally in
a final list and all selection changes modify this single list instance.
This instance is always set to the ListModel in
UiCommonEditorVisitor::setInModel. However, ListModel::setSelectedItems
fires SelectedItemsChangedEvent only if the incoming list of selected
items is different to the previous one which is never true. As a result,
the event is never fired and it is not possible to listen to the
selection events and react accordingly.

This patch creates a new array object on a change, causing the event to
raise.

Change-Id: Ice428f88a15d189ffed576d38fec4fe8c0052d2b
Bug-Url: https://bugzilla.redhat.com/1964496
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>