core: change description of the snapshot of clone-vm
This patch changes the description of the snapshot that is automatically
created during CloneVm from "Auto-generated for Clone Live VM" to
"Auto-generated for Clone VM" since the snapshot is created, by design,
also when cloning a VM that is down.

The constant that holds that description is also renamed as LCV part
that stands for Live Clone VM is incorrect.

Change-Id: I24b5daf0f897b6f838b9c22cfbfbd37d5dd7e449
Signed-off-by: Arik Hadas <ahadas@redhat.com>