restapi - 'max' parameter problem
GET ../api/disks using a search query and 'max' parameter
doesn't work, this patch fixes the probelm.

NOTE: for systems with large amounts of entities this fix
may be heavy in terms of performance (only for the case
of filtered-query in combination with search-query and 'max')

NOTE: while not mentioned in the bugzilla, the same problem
is expected to happen for all non-admin requests which combine
a search predicate and use of 'max'. This patch should fix
those as well

Bug-Url: https://bugzilla.redhat.com/1962177
Signed-off-by: Ori Liel <oliel@redhat.com>
Change-Id: I2631fa81fe81d3d57d85357e2175f27dc258c092